refactor: remove PolarionBaselineExecutor and related test, switch to generic service PolarionBaselineExecutor
